Cloud-Computations/requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.0
numpy==1.20.2
pandas==1.2.3
patsy==0.5.1
Pillow==8.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.16.0
pytz==2021.1
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
sklearn==0.0
statsmodels==0.12.2
threadpoolctl==2.1.0
urllib3==1.26.4
